a little more cleanup, better error reporting
